requirements.txt
streamlit>=1.32
opencv-python-headless>=4.8
numpy>=1.24
pillow>=10.0
